refactor: 💡 migrate credential-store off fragments
