--- addons/api/tests/unit/models/fragment-credential-store-attributes-test.js
@@ -1,13 +1,0 @@
-import { module, test } from 'qunit';
-import { setupTest } from 'ember-qunit';
-
-module('Unit | Model | fragment credential store attributes', function (hooks) {
-  setupTest(hooks);
-
-  // Replace this with your real tests.
-  test('it exists', function (assert) {
-    let store = this.owner.lookup('service:store');
-    let model = store.createRecord('fragment-credential-store-attributes', {});
-    assert.ok(model);
-  });
-});
